requirements.txt
azure-cognitiveservices-nspkg==2.0.0
azure-cognitiveservices-search-newssearch==1.0.0
azure-cognitiveservices-search-nspkg==2.0.0
azure-common==1.1.15
azure-nspkg==2.0.0
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
djangorestframework==3.8.2
idna==2.7
isodate==0.6.0
msrest==0.5.5
oauthlib==2.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
